Introducing Repository Unit Tests

+++ src/main/java/com/tas/repository/OrganizationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.tas.bean.Organization;
 
-public interface OrganizationRepository extends JpaRepository<Organization, Integer>{
+public interface OrganizationRepository extends JpaRepository<Organization, Long>{
 
 }
